Prompt history (#45)
Add a simple prompt history management in session.
Trigger it with up/down arrows from first/last lines of input to recall
your previous or next prompts in the current session.
Also added multiline input management in the terminal (Alt+Enter as in
claude code, but we may want to align with the web and have cmd+enter
submit...)

---------

Co-authored-by: vincent-audibert-whoz <vincent.audibert@whoz.com>